google_cpm: use ->cp_hold for wireless and wired charging
The select logic based on demand will place an hold on CP charging
when battery demand is lowered byt thermals and charging switch
from CP to the default charger.

The hold debounces re-entry to CP charging and is reset from only
a change to the thermal limit (legacy or dissipation based).

Bug: 198700183
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Ifc6bc067cff32087a050ad1eb69a161c09040b79