OWNERS: add more owners
Since this branch is managed by the Android Systems Team, we need to
have some ownership to approve changes when necessary. In particular, we
will be enabling this project in presubmit for the Android Common Kernel
and need the permissions to quickly fix build issues that block our
presubmit if they arise. Regular bug fixes that are not presubmit
blocking, should still be approved by the driver owners.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ibacd1e280fc40665d3a24ba0e61f79266b6efddd