google_cpm: reset start time on stop
There is a race between the stop of CP charging for thermals and the
restart that causes PPS to be disabled for the session. This change
makes sure that the start of the session is initialized and cleared
properly.

Bug: 231656227
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: If341d6b4b31a71c2a1d2f57cc37bdc14e2252e8a
(cherry picked from commit 82d83103f2cf389cb389bfdc9e40305e0261dd78)