max77759_charger: use of cached wcin_input_supend
The callback for max77750 charge mode uses the cached value of
wcin_input_suspend. max77759_chgin_input_suspend() set the
cached value to the current one before voting on the charger
mode to make sure that the callback uses the right value.

This will unmask the WCIN input when the themal level drops from
critical and will resume wireless charging as soon as possible.

Bug: 231656227
Signed-off-by: AleX Pelosi <apelosi@google.com>
Change-Id: Ia626e8f39b2d892758a07a5baee95dbbae43769e