bms/misc: Remove default =m in the Makefile
Now that we have Kconfig support, we can drop setting =m unconditionally
in the Makefile and rely on the defconfig to enable/disable this config.
I believe there are issues still related to not using #if
IS_ENABLED(...) for this config. So the build will break if it's
disabled, but that is not introduced by this patch.

Bug: 231997546
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ic2aaa2ccc11455bc982dbad0e22c37c74ad62b17