max1720x_battery: move record FG data proccess to cycle check function
The original design was to check the value and save when the adapter was disconnected. However, the charger is notified of disconnection during the shutdown process, which may result in incomplete data saved to the battery eeprom.

Consider vendor design doc., trigger condition related to FG cycles.
Move to get_cycle_count function to check when read cycle data

Bug: 251353728
Signed-off-by: Jenny Ho <hsiufangho@google.com>
Change-Id: I1b4e0742f10e8361c120ebca2a96258ffd56bbc0