google_battery: Switch to new of thermal API
The thermal OF code has a new API allowing to migrate the OF
initialization to a simpler approach. The ops are no longer device
tree specific and are the generic ones provided by the core code.

Convert the ops to the thermal_zone_device_ops format and use the new
API to register the thermal zone with these generic ops. This is
required starting in kernel 6.1.

Bug: 251537418
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I813ac6652a0677cb8c19af571ce078ccd9bf719d