max77779_i2cm: I2C bus driver for I2CM Bridge
The I2CM bridge provides a register interface to a remote bus.  This
driver exposes an I2C bus that allows remote peripherals to (in most
cases) work with out modification.  The one limitation is that the
bridge only support the SMBUS subset of i2c.  Transactions with data
larger than 32 bytes need to be split up by the caller.  Regmap will
do this transparently.

Test: Any i2c transactions
Bug: 274594312
Change-Id: I6fe495c9732a2d17a16dfc413f4a93dade7410c3
Signed-off-by: Jim Wylder <jwylder@google.com>