max77779_pmic: PMIC Subsystem Functions Top Level
PMIC subsystem of MAX77779 provides a top level interrupts and control
over a group of pins. Since the functionality for each requires a fair
amount of structural boilerplate code, the logic is divided into
multiple files.  This top level file handles the i2c communication with
the registers in the PMIC subsystem.

Test: Run with Sequoia PMIC
Bug: 274594312
Change-Id: I09159679f5730b6a614ba81b1dc4ffb52ab2b319
Signed-off-by: Jim Wylder <jwylder@google.com>