max77779_pmic: Add IRQ Multiplexing
PMIC subsystem of MAX77779 provides a top level interrupts and control
over a group of pins. Since the functionality for each requires a fair
amount of structural boilerplate code, the logic is divided into
multiple files.

This commit adds the an interrupt controller to multiplex a single IRQ
line to the interrupt sources as listed int the INTSRC_STS register.
The irq functionality is kicked off by the main max77779_pmic module.
This allows the "device" needed for the register access functions to be
passed to the probe.

Test: Run with Sequoia PMIC
Bug: 274594312
Change-Id: I9c2c16b8073818696c3619d153f5dacf2246352f
Signed-off-by: Jim Wylder <jwylder@google.com>