gbms: logbuffer register: fix the error path on alloc failure
In logbuffer_register(), the label 'free_instance' is responsible for
freeing the instance on error, and returning an error value.  Setting
the instance value to an error value before jumping to 'free_instance'
is unnecessary and results in memory leak, and it is also undefined
behavior to pass the error value to kfree().

Don't change the value of the instance before jumping to the error path
(i.e., 'free_instance').

Bug: 297950261
Change-Id: I19157d08fbff6ed63806409a673deba2edef0cbf
Signed-off-by: Necip Fazil Yildiran <necip@google.com>
(cherry picked from commit e77bce56dc5ac10d1b369667f8c03b00c6142361)