max77779: Move calls to irq_set_irq_wake to unlock code
Calls to irq_set_irq_wake() call the bus_lock() for the irq_chip, which
takes a mutex.  It then takes a spinlock.  So any further calls to
irq_set_irq_wake() that take a mutex can't be made.  In general, work
that could lock is best removed from all of the irq functions and done
in the driver's unlock.

Test: Build with --config=debug_locking
Bug: 322564428
Change-Id: Iedb4c7fb86c393a79cdde327a2aca23094289dea
Signed-off-by: Jim Wylder <jwylder@google.com>