max77779: sgpio: Support Unmask
For max77779, masking/unmasking of an irq is handled by setting the
appropriate trigger type value in the IRQ_SEL field for the gpio.
Previous version only updated the IRQ_SEL field when unmasking or
directly changing the trigger type.

Update the unlock code to update the IRQ_SEL when there is any change to
either the mask or the tigger type.  NOTE: the new behavior for the
case, where no trigger type is ever set but the irq is unmasked will be
to basically leave the trigger disabled.

Test: Bench test
Bug: 318396700
Change-Id: Ifc5d8d04cf5a557732a8ab0a8c336ce098538de4
Signed-off-by: Jim Wylder <jwylder@google.com>