max77779_vimon: buffer read through delayed work
Upon vimon IRQ triggering, the vimon driver shall collect the available
data from the buffer starting at rfap register, for a count of rsc sets
of voltage and current readings. This data collection is performed
through a delayed work. The collected data will be made available by a
exported function. Additional clean up work is done to allow proper
functionality of the driver:
- removing the IRQ_NONE returns from max77779_vimon_init_fs which
presently cause the IRQ to be repeatedly called.
- Checking the INT_IRQ_STS against the BVIM_MASK register is removed,
as the implemented logic is assuming if BVIM_MASK is asserted, the IRQ
will be registered. Instead, the hardware will not trigger an IRQ if
the BVIM_MASK bit is asserted.
- Some vimon_lock mutex_lock are removed, as they were only guarding an
enum assignment.
- In max77779_vimon_init, the config mask to stop data collection upon
batoilo1 and batoilo2 are removed, and instead the data collection is
allowed to continue until reaching smpl_cnt value.

Bug: 296133383
Change-Id: Ieabe859cf069557a97f45270cdd4979f0d1fc469
Signed-off-by: Hiroshi Akiyama <hiroshiakiyama@google.com>