google_batery: clear shutdown flag after successfully booting to OS
When the device boots into OS from critical shutdown,
shutdown flag will be stored in a variable. If there is
no manual shutdown or reboot, shutdown flag will not
be cleared, and will be written to the storage when the
device encounter critical shutdown again.

Fix it by clearing shutdown flag when the charger is
plugged in and the SOC is greater than or equal to 5%.

Bug: 339780394
Change-Id: Ie0ad9065f6ea27be9849437a0298ec83df31985d
Signed-off-by: Jack Wu <wjack@google.com>